+++ api/app/clients/prompts/index.js
@@ -1,0 +1,9 @@
+const instructions = require('./instructions');
+const titlePrompts = require('./titlePrompts');
+const refinePrompts = require('./refinePrompts');
+
+module.exports = {
+  ...refinePrompts,
+  ...instructions,
+  ...titlePrompts,
+};
